Сделана вьюмодель и фейковый репозиторий.
