+++ app/src/main/java/profitsw2000/diffapps/domain/Repository.kt
@@ -1,0 +1,10 @@
+package profitsw2000.diffapps.domain
+
+import io.reactivex.rxjava3.core.Single
+import profitsw2000.diffapps.model.DayBloodPressure
+
+interface Repository {
+
+    fun getAllBPMeasurementResults() : Single<List<DayBloodPressure>>
+
+}
