+++ app/src/main/java/profitsw2000/diffapps/presentation/view/adapters/BloodPressureResultsAdapter.kt
@@ -1,6 +1,5 @@
 package profitsw2000.diffapps.presentation.view.adapters
 
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
